# swagger_server/models/prereq.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from swagger_server.models.base_model_ import Model
from swagger_server import util


class Prereq(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """

    def __init__(self, prereq_group_id: int=None, class_dept: str=None, class_number: int=None):  # noqa: E501
        """Prereq - a model defined in Swagger

        :param prereq_group_id: The prereq_group_id of this Prereq.  # noqa: E501
        :type prereq_group_id: int
        :param class_dept: The class_dept of this Prereq.  # noqa: E501
        :type class_dept: str
        :param class_number: The class_number of this Prereq.  # noqa: E501
        :type class_number: int
        """
        self.swagger_types = {
            'prereq_group_id': int,
            'class_dept': str,
            'class_number': int
        }

        self.attribute_map = {
            'prereq_group_id': 'prereq_group_id',
            'class_dept': 'class_dept',
            'class_number': 'class_number'
        }

        self._prereq_group_id = prereq_group_id
        self._class_dept = class_dept
        self._class_number = class_number

    @classmethod
    def from_dict(cls, dikt) -> 'Prereq':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The Prereq of this Prereq.  # noqa: E501
        :rtype: Prereq
        """
        return util.deserialize_model(dikt, cls)

    @property
    def prereq_group_id(self) -> int:
        """Gets the prereq_group_id of this Prereq.


        :return: The prereq_group_id of this Prereq.
        :rtype: int
        """
        return self._prereq_group_id

    @prereq_group_id.setter
    def prereq_group_id(self, prereq_group_id: int):
        """Sets the prereq_group_id of this Prereq.


        :param prereq_group_id: The prereq_group_id of this Prereq.
        :type prereq_group_id: int
        """
        if prereq_group_id is not None and prereq_group_id < 1:  # noqa: E501
            raise ValueError("Invalid value for `prereq_group_id`, must be a value greater than or equal to `1`")  # noqa: E501

        self._prereq_group_id = prereq_group_id

    @property
    def class_dept(self) -> str:
        """Gets the class_dept of this Prereq.


        :return: The class_dept of this Prereq.
        :rtype: str
        """
        return self._class_dept

    @class_dept.setter
    def class_dept(self, class_dept: str):
        """Sets the class_dept of this Prereq.


        :param class_dept: The class_dept of this Prereq.
        :type class_dept: str
        """
        if class_dept is not None and len(class_dept) > 5:
            raise ValueError("Invalid value for `class_dept`, length must be less than or equal to `5`")  # noqa: E501
        if class_dept is not None and len(class_dept) < 2:
            raise ValueError("Invalid value for `class_dept`, length must be greater than or equal to `2`")  # noqa: E501

        self._class_dept = class_dept

    @property
    def class_number(self) -> int:
        """Gets the class_number of this Prereq.


        :return: The class_number of this Prereq.
        :rtype: int
        """
        return self._class_number

    @class_number.setter
    def class_number(self, class_number: int):
        """Sets the class_number of this Prereq.


        :param class_number: The class_number of this Prereq.
        :type class_number: int
        """
        if class_number is not None and class_number > 9999:  # noqa: E501
            raise ValueError("Invalid value for `class_number`, must be a value less than or equal to `9999`")  # noqa: E501
        if class_number is not None and class_number < 0:  # noqa: E501
            raise ValueError("Invalid value for `class_number`, must be a value greater than or equal to `0`")  # noqa: E501

        self._class_number = class_number