# swagger_server/swagger/swagger.yaml
...
swagger: "2.0"
info:
  description: "This a API for the ClassDeck Project. Used for communication between\
    \ front-end and back-end.\n"
  version: "1.0.0"
  title: "ClassDeck Project"
  termsOfService: "http://swagger.io/terms/"
  contact:
    email: "kvitca.p@husky.neu.edu"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "virtserver.swaggerhub.com"
basePath: "/pablokvitca/classdeck-api/1.0.0"
...
schemes:
- "https"
- "http"
paths:
...
  /prereq/class:
    post:
      tags:
      - "prereq"
      summary: "Add a prereq to the classdeck"
      operationId: "add_prereq_class"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Prereq object that needs to be added to the system"
        required: true
        schema:
          $ref: "#/definitions/Prereq"
      responses:
        405:
          description: "Invalid input"
      security:
      - internal_auth: []
      x-swagger-router-controller: "swagger_server.controllers.prereq_controller"
    put:
      tags:
      - "prereq"
      summary: "Update an existing prereq"
      operationId: "update_prereq_class"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Prereq object that needs to be updated in the system"
        required: true
        schema:
          $ref: "#/definitions/Prereq"
      responses:
        400:
          description: "Invalid ID supplied"
        404:
          description: "Prereq not found"
        405:
          description: "Validation exception"
      security:
      - internal_auth: []
      x-swagger-router-controller: "swagger_server.controllers.prereq_controller"
  /prereq/class/{prereq_group_id}/{class_dept}/{class_number}:
    get:
      tags:
      - "prereq"
      summary: "Find prereq by ID"
      description: "Returns a single prereq"
      operationId: "get_prereq_class_by_id"
      produces:
      - "application/json"
      parameters:
      - name: "prereq_group_id"
        in: "path"
        description: "ID of prereq group related to the prereq to return"
        required: true
        type: "integer"
        minimum: 0
      - name: "class_dept"
        in: "path"
        description: "short code of the department of the class related to the prereq\
          \ to return"
        required: true
        type: "string"
        maxLength: 5
        minLength: 2
      - name: "class_number"
        in: "path"
        description: "number of the class related to the prereq to return"
        required: true
        type: "integer"
        maximum: 9999
        minimum: 0
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/Prereq"
        400:
          description: "Invalid ID supplied"
        404:
          description: "Prereq not found"
      security:
      - internal_auth: []
      x-swagger-router-controller: "swagger_server.controllers.prereq_controller"
    delete:
      tags:
      - "prereq"
      summary: "Deletes a prereq"
      operationId: "delete_prereq_class"
      produces:
      - "application/json"
      parameters:
      - name: "prereq_group_id"
        in: "path"
        description: "ID of prereq group related to the prereq to return"
        required: true
        type: "integer"
        minimum: 0
      - name: "class_dept"
        in: "path"
        description: "short code of the department of the class related to the prereq\
          \ to return"
        required: true
        type: "string"
        maxLength: 5
        minLength: 2
      - name: "class_number"
        in: "path"
        description: "number of the class related to the prereq to return"
        required: true
        type: "integer"
        maximum: 9999
        minimum: 0
      responses:
        400:
          description: "Invalid ID supplied"
        404:
          description: "Prereq not found"
      security:
      - internal_auth: []
      x-swagger-router-controller: "swagger_server.controllers.prereq_controller"
  /prereq/degree:
    post:
      tags:
      - "prereq"
      summary: "Add a prereq to the classdeck"
      operationId: "add_prereq_degree"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Prereq object that needs to be added to the system"
        required: true
        schema:
          $ref: "#/definitions/Prereq"
      responses:
        405:
          description: "Invalid input"
      security:
      - internal_auth: []
      x-swagger-router-controller: "swagger_server.controllers.prereq_controller"
    put:
      tags:
      - "prereq"
      summary: "Update an existing prereq"
      operationId: "update_prereq_degree"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Prereq object that needs to be updated in the system"
        required: true
        schema:
          $ref: "#/definitions/Prereq"
      responses:
        400:
          description: "Invalid ID supplied"
        404:
          description: "Prereq not found"
        405:
          description: "Validation exception"
      security:
      - internal_auth: []
      x-swagger-router-controller: "swagger_server.controllers.prereq_controller"
  /prereq/degree/{prereq_group_id}/{class_dept}/{class_number}:
    get:
      tags:
      - "prereq"
      summary: "Find prereq by ID"
      description: "Returns a single prereq"
      operationId: "get_prereq_degree_by_id"
      produces:
      - "application/json"
      parameters:
      - name: "prereq_group_id"
        in: "path"
        description: "ID of prereq group related to the prereq to return"
        required: true
        type: "integer"
        minimum: 0
      - name: "class_dept"
        in: "path"
        description: "short code of the department of the class related to the prereq\
          \ to return"
        required: true
        type: "string"
        maxLength: 5
        minLength: 2
      - name: "class_number"
        in: "path"
        description: "number of the class related to the prereq to return"
        required: true
        type: "integer"
        maximum: 9999
        minimum: 0
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/Prereq"
        400:
          description: "Invalid ID supplied"
        404:
          description: "Prereq not found"
      security:
      - internal_auth: []
      x-swagger-router-controller: "swagger_server.controllers.prereq_controller"
    delete:
      tags:
      - "prereq"
      summary: "Deletes a prereq"
      operationId: "delete_prereq_degree"
      produces:
      - "application/json"
      parameters:
      - name: "prereq_group_id"
        in: "path"
        description: "ID of prereq group related to the prereq to return"
        required: true
        type: "integer"
        minimum: 0
      - name: "class_dept"
        in: "path"
        description: "short code of the department of the class related to the prereq\
          \ to return"
        required: true
        type: "string"
        maxLength: 5
        minLength: 2
      - name: "class_number"
        in: "path"
        description: "number of the class related to the prereq to return"
        required: true
        type: "integer"
        maximum: 9999
        minimum: 0
      responses:
        400:
          description: "Invalid ID supplied"
        404:
          description: "Prereq not found"
      security:
      - internal_auth: []
      x-swagger-router-controller: "swagger_server.controllers.prereq_controller"
...
securityDefinitions:
  classdeck_auth:
    type: "oauth2"
    authorizationUrl: "http://classdeck.swagger.io/oauth/dialog"
    flow: "implicit"
    scopes:
      write: "write anything on your account"
      read: "read anything on your account"
  internal_auth:
    type: "apiKey"
    name: "internal_auth"
    in: "header"
definitions:
...
  Prereq:
    type: "object"
    properties:
      prereq_group_id:
        type: "integer"
        minimum: 1
      class_dept:
        type: "string"
        minLength: 2
        maxLength: 5
      class_number:
        type: "integer"
        minimum: 0
        maximum: 9999
    example:
      prereq_group_id: 1
      class_dept: "XX"
      class_number: 9999
...